@@ -2,28 +2,25 @@
     "name": "FKM",
     "private": true,
     "dependencies": {
-        "react": "18.0.0",
-        "react-dom": "18.0.0",
-        "react-router-dom": "6.3.0",
-        "@mui/material": "5.6.0",
-        "@emotion/react": "11.9.0",
-        "@emotion/styled": "11.8.1"
+        "react": "18.2.0",
+        "react-dom": "18.2.0",
+        "react-router-dom": "6.8.0",
+        "@mui/material": "5.11.7",
+        "@mui/icons-material": "5.11.0",
+        "@emotion/react": "11.10.5",
+        "@emotion/styled": "11.10.5"
     },
     "devDependencies": {
-        "@types/react": "17.0.43",
-        "@types/react-dom": "17.0.14",
+        "@types/react": "18.0.27",
+        "@types/react-dom": "18.0.10",
         "@types/react-router-dom": "5.3.3",
-        "css-loader": "6.7.1",
-        "fork-ts-checker-webpack-plugin": "7.2.3",
-        "sass-embedded": "1.49.11",
-        "sass-loader": "12.6.0",
-        "source-map-loader": "3.0.1",
+        "css-loader": "6.7.3",    
         "style-loader": "3.3.1",
-        "ts-loader": "9.2.8",
-        "typescript": "4.6.3",
-        "typescript-plugin-proactive-code-fixes": "0.0.18",
-        "webpack": "5.71.0",
-        "webpack-cli": "4.9.2"
+        "ts-loader": "9.4.2",
+        "typescript": "4.9.5", 
+        "webpack": "5.75.0",
+        "webpack-cli": "5.0.1",
+        "tslib": "2.5.0"
     },
     "scripts": {
         "dev": "webpack --mode \"development\" --devtool \"eval-source-map\" --progress --watch",
