@@ -1,4 +1,5 @@
 using DoItTest.Site;
+using DoItTest.Site.Infrastructure.Filters;
 using DoItTest.Tools.Json;
 
 WebApplicationBuilder builder = WebApplication.CreateBuilder(args);
@@ -7,7 +8,7 @@ builder.Initialize();
 
 builder.Services.AddControllersWithViews(mvcOptions =>
 {
-
+    mvcOptions.Filters.Add<IsAuthorizedFilter>();
 }).AddJsonOptions(options =>
 {
     options.JsonSerializerOptions.AddJsonSettings();
