@@ -1,10 +1,15 @@
-﻿@{Layout = null;}
+﻿@model DoItTest.Site.Infrastructure.ReactApp;
+@{Layout = null;}
 <!DOCTYPE html>
 <html>
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Test</title>
+    <script>
+        var systemUser = @Html.Raw(Json.Serialize(Model.SystemUser));
+        var sidebarLinksTree =  @Html.Raw(Json.Serialize(Model.SidebarLinksTree));
+    </script>
 </head>
 <body>
     <div id="app"></div>
